refactor: moved operators into object_ module
